% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymatreg.R
\name{mymatreg}
\alias{mymatreg}
\title{mymatreg}
\usage{
mymatreg(obj)
}
\arguments{
\item{obj}{an object from an `lm()`}
}
\value{
an invisible list to call including: summary, data frame, characteristic & response matrices, beta estimates, RSS, ssq, s, fitted values, residuals, leverage values, and a resiudal plot.
}
\description{
A function which returns many great things from matrix algebra for regression
}
\examples{
\dontrun{mymatreg(ylm)}
}
